Move more logic into Prometheus metric data getter
move the logic for removing undefined metrics into the base metrics data
getter for prometheus

this will make it easier to add more prometheus metric data getters
